Fix lint on mac os
